+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `place_perp_order_with_max_ts`
 - Fix precision loss in `ZoDexMarket::lots_to_price`
 - BREAKING: Change accounts passed to `cache_oracle`
+- BREAKING: Bump `anchor-lang` to `v0.25.0`
 
 ## [0.5.0] - 2022-04-25